blast-helper util added for helper functions, renamings
