--- packages/blast-utilities/delay.js
@@ -1,10 +1,0 @@
-
-async function delay(n) {
-  return new Promise(function(resolve) {
-    setTimeout(resolve, n * 1000)
-  })
-}
-
-module.exports = {
-  delay: delay
-}
