+++ packages/blast-utilities/fs-utils.js
@@ -3,12 +3,12 @@ const { getProjectRootPath } = require('./package-info')
 const BlastError = require('./blast-error')
 
 function saveAccounts(accounts) {
-  const pkgRoot = getProjectRootPath()
+  const projectRoot = getProjectRootPath()
   const parsed = JSON.parse(accounts)
   try {
-    fs.writeFileSync(`${pkgRoot}/accounts.json`, JSON.stringify(parsed, 0, 4))
+    fs.writeFileSync(`${projectRoot}/accounts.json`, JSON.stringify(parsed, 0, 4))
   } catch (error) {
-    throw new BlastError(`Failed to create file at ${pkgRoot}/additional-accounts.json with error: ${error}`)
+    throw new BlastError(`Failed to create file at ${projectRoot}/additional-accounts.json with error: ${error}`)
   }
 }
