+++ template/blast.config.js
@@ -3,7 +3,7 @@ module.exports.config = {
   name: 'template',
   network: 'cudos',
   defaultAccount: 'account1',
-  additionalCustomAccounts: 0, // default is 10
+  additionalAccounts: 0,
   customAccountBalances: '1000000000000000000',
   gasPrice: '250acudos',
   endpoint: 'http://localhost:26657'
